Fix TestDestroySetsEncryptionsalt test and resulting bug (#15432)
<!--- 
Thanks so much for your contribution! If this is your first time
contributing, please ensure that you have read the
[CONTRIBUTING](https://github.com/pulumi/pulumi/blob/master/CONTRIBUTING.md)
documentation.
-->

# Description

<!--- Please include a summary of the change and which issue is fixed.
Please also include relevant motivation and context. -->

While working on some other secret manager changes I found that
`TestDestroySetsEncryptedkey` wasn't actually fully testing what we
thought it was.

Firstly it was a bad name, we're checking for `encryptionsalt` being set
not `encryptedkey`. But more importantly it wasn't checking that the
salt stayed the same.

Turned out `destroy` was loading the stack config, seeing no
`encryptionsalt` and so new'ing up a brand new passphrase secret manager
and state and then saving that to the stack config.

This is now fixed that the test asserts that the salt is exactly what's
expected and I've fixed up the engine code to do this correctly.

## Checklist

- [x] I have run `make tidy` to update any new dependencies
- [x] I have run `make lint` to verify my code passes the lint check
  - [x] I have formatted my code using `gofumpt`

<!--- Please provide details if the checkbox below is to be left
unchecked. -->
- [x] I have added tests that prove my fix is effective or that my
feature works
<!--- 
User-facing changes require a CHANGELOG entry.
-->
- [x] I have run `make changelog` and committed the
`changelog/pending/<file>` documenting my change
<!--
If the change(s) in this PR is a modification of an existing call to the
Pulumi Cloud,
then the service should honor older versions of the CLI where this
change would not exist.
You must then bump the API version in
/pkg/backend/httpstate/client/api.go, as well as add
it to the service.
-->
- [ ] Yes, there are changes in this PR that warrants bumping the Pulumi
Cloud API version
<!-- @Pulumi employees: If yes, you must submit corresponding changes in
the service repo. -->