Add test for legacy style names (#15384)
Long term I want to introduce a new snake_case based naming system to
schemas. There's an internal design doc about this at
https://docs.google.com/document/d/1ns07htpLjw0dJPn5p7TBPVULvwwYc1Pl7l1c1jdGXoY/edit.

As part of that we get the nice feature that _nearly_ all current names
are either valid and the same in both system (e.g. "foo"), or they look
different in the current system (e.g. "fooBar"). But there are a handful
of odd cases where we have providers using names that look like they
should be new style snake case names (e.g. kubernetes has a load of
"x_kubernetes_" properties).

To provide a way to ensure k8s can stay on the current name generation
(even though it gives pretty odd results) this change adds a test with a
schema demonstrating a load of snake_case names. When we do add the new
naming system this test _should not diff_, excepting possibly setting an
option if the new naming system is opt-out rather than opt-in.