Enable program test to install dev releases of SDKs (#15387)
# Description

Add a new option `InstallDevReleases` to `ProgramTestOptions`, which
will tell ProgramTest to upgrade each pulumi dependency to the latest
dev release. This includes the pulumi/pulumi SDKs and the provider SDKs.
The idea is to use this in the `pulumi/examples` and `pulumi/templates`
repositories to get pre release testing there.

Fixes https://github.com/pulumi/pulumi/issues/15389

## Checklist

- [x] I have run `make tidy` to update any new dependencies
- [x] I have run `make lint` to verify my code passes the lint check
  - [x] I have formatted my code using `gofumpt`

<!--- Please provide details if the checkbox below is to be left
unchecked. -->
- [ ] I have added tests that prove my fix is effective or that my
feature works
<!--- 
User-facing changes require a CHANGELOG entry.
-->
- [x] I have run `make changelog` and committed the
`changelog/pending/<file>` documenting my change
<!--
If the change(s) in this PR is a modification of an existing call to the
Pulumi Cloud,
then the service should honor older versions of the CLI where this
change would not exist.
You must then bump the API version in
/pkg/backend/httpstate/client/api.go, as well as add
it to the service.
-->
- [ ] Yes, there are changes in this PR that warrants bumping the Pulumi
Cloud API version
<!-- @Pulumi employees: If yes, you must submit corresponding changes in
the service repo. -->