convert-trace handles root spans with no TimespanEvent
If a span doesn't have a TimespanEvent then determine the start and
end times from the sub spans (if any). This fixes an occational issue
where one of the root spans doesn't have the TimespanEvent set which
causes the entire OTEL trace conversion to fail

Also ensure that InstrumentationLibrary and InstrumentationScope return
the same object.