Exit no-op `pulumi state edit`
The current behavior of `pulumi state edit` when the edited file is unchanged is to prompt
the user to `accept`, `edit`, `reset` or `cancel`. Since the file is unchanged, `accept`
and `cancel` do the same thing. `edit` (again) is the same as `reset` which is the same as
running the command from scratch.

The feature would be much easier to use if a no-op edit simply exited. If the user wants
to edit again, they can just run `!!` to repeat the command.