push post-release changelog and go.mod updates straight to master
Post release, we need to update the changelog (clear out all the
pending entries), and update the go.mod in pkg, and also tag the
latest version of that.  Currently we create a new PR for that.
However since we are using squash merges, and need to do the tagging
while creating the PR, we're tagging a commit that doesn't end up on
the main branch.  This means the Go tooling doesn't work properly with
pkg.  See also https://github.com/pulumi/pulumi/issues/15458.

There is nothing really we are doing with that PR other than merging
it (in fact queue-merge is set to true, so it should be queued and
merged automatically, but that doesn't seem to work).

Because of that, we can just tag the release, and push it straight to
the main branch instead.  This removes one human step from the release
process and makes the Go tooling work correctly.

The only potential downsides of this change are:
- a small race condition, if a commit gets merged just before we do
the rebase we might end up failing that and the release workflow
fails.  This is very unlikely, since the release process should be
relatively quick.
- we need to adjust the branch protection rules for pulumi-bot so it's
able to push to the main branch directly.  This should not be a
problem, since it's an automated tool, so all of its potential
activity has been reviewed by humans beforehand.