Decouple persist and display events
This removes a scenario where events could not be persisted to the cloud
because they were waiting on the same event being displayed

The primary change, however, is to stop rendering the tree every time
a row is updated, instead, theis renders when the display actually
happens in the the `frame` call. The renderer instead simply marks
itself as dirty in the `rowUpdated`, `tick`, `systemMessage` and `done`
methods and relies on the frame being redrawn on a 60Hz timer (the
`done` method calls `frame` explicitly)

This requires a minor refactor to ensure that the display object is
available in the frame method

Because the treeRenderer is calling back into the ProgressDisplay object
in a goroutine, the ProgressDisplay object needs to be thread safe, so
a read-write mutex is added to protect the `eventUrnToResourceRow` map.
The unused `urnToID` map was removed in passing.