enable automatic plugin acquisition in python acceptance test
This test was disabled on MacOS due to what looks like a failure to
acquire a plugin.  This probably worked accidentally (and is working
accidentally on other platforms) because presumably the plugin gets
acquired by another test and is thus available.  We should however not
rely on that, end enable plugin acquisition where necessary.  With
that we also fix the test on MacOS and can unskip it.