acquire plugin in python acceptance test (#15542)
This test was disabled on MacOS due to what looks like a failure to
acquire a plugin. This probably worked accidentally (and is working
accidentally on other platforms) because presumably the plugin gets
acquired by another test and is thus available. We should however not
rely on that, end enable plugin acquisition where necessary. With that
we also fix the test on MacOS and can unskip it.

This addresses part of https://github.com/pulumi/pulumi/issues/15240.
Note that I don't think it fixes the flakyness, looking back through the
test logs I think I was conflating two separate things here.