Fix codepaths integration tests
Deps might include more than just the direct dependencies, but the
precise results depend on how the packages are hoisted within
node_modules. This can change based on the version of the package
manager and the dependencies of @pulumi/pulumi.

For example for this nesting:

node_modules
└─┬ semver
    └── lru-cache

deps only includes `node_modules/semver`. However if they are siblings,
it will include both `node_modules/semver` and `node_modules/lru-cache`.

We only assert that the direct dependencies are included, which are
guaranteed to be stable.