show all paths when we have multipe local policy packs with the same name
It is possible for multiple local policy packs to have the same name.
When there are policy violation, we currently only show one of the
paths for those policy packs (as we consider them the same policy pack
because it has the same name).

Display both of the paths in this case to avoid any confusion.  Also
sort the names, so the test we have for this is deterministic.