don't run conformance test tests in parallel (#15584)
Currently conformance tests can't be run in parallel because the engine
will internally chdir, and have some global state. This is tracked in
https://github.com/pulumi/pulumi/issues/13945. This means also the tests
testing the pulumi-test-language program can't run in parallel, or they
will be flaky.

Make them execute serially and link to the issue mentioned above to
de-flake them.

Fixes #15580